+++ language-server/Shared.AnZwDev.ALTools/WorkspaceCommands/AddReferencedTablesPermissionsWorkspaceCommand.cs
@@ -14,6 +14,9 @@ namespace AnZwDev.ALTools.WorkspaceCommands
 
     public class AddReferencedTablesPermissionsWorkspaceCommand : SemanticModelSyntaxRewriterWorkspaceCommand<AddReferencedTablesPermissionsSyntaxRewriter>
     {
+
+        public static string SortPropertiesParameterName = "sortProperties";
+
         public AddReferencedTablesPermissionsWorkspaceCommand(ALDevToolsServer alDevToolsServer) : base(alDevToolsServer, "addReferencedTablesPermissions")
         {
         }
@@ -23,6 +26,13 @@ namespace AnZwDev.ALTools.WorkspaceCommands
             //disable format because default formatter keeps all permissions in a single line
             return node;
         }
+
+        protected override void SetParameters(SyntaxTree syntaxTree, SyntaxNode node, SemanticModel semanticModel, ALProject project, TextSpan span, Dictionary<string, string> parameters)
+        {
+            base.SetParameters(syntaxTree, node, semanticModel, project, span, parameters);
+            this.SyntaxRewriter.SortProperties = parameters.GetBoolValue(SortPropertiesParameterName);
+        }
+
 
     }