+++ language-server/Shared.AnZwDev.ALTools/WorkspaceCommands/SetPageFieldToolTipWorkspaceCommand.cs
@@ -13,6 +13,7 @@ namespace AnZwDev.ALTools.WorkspaceCommands
     {
         public static string ToolTipParameterName = "toolTip";
         public static string CommentParameterName = "comment";
+        public static string SortPropertiesParameterName = "sortProperties";
 
         public SetPageFieldToolTipWorkspaceCommand(ALDevToolsServer alDevToolsServer) : base(alDevToolsServer, "setPageFieldToolTip")
         {
@@ -23,6 +24,7 @@ namespace AnZwDev.ALTools.WorkspaceCommands
             base.SetParameters(sourceCode, project, filePath, span, parameters);
             this.SyntaxRewriter.ToolTip = parameters.GetStringValue(ToolTipParameterName);
             this.SyntaxRewriter.Comment = parameters.GetStringValue(CommentParameterName);
+            this.SyntaxRewriter.SortProperties = parameters.GetBoolValue(SortPropertiesParameterName);
         }
 
     }