+++ vscode-extension/src/alsyntaxmodifiers/objectCaptionsModifier.ts
@@ -1,3 +1,4 @@
+import * as vscode from 'vscode';
 import { DevToolsExtensionContext } from "../devToolsExtensionContext";
 import { WorkspaceCommandSyntaxModifier } from "./workspaceCommandSyntaxModifier";
 
@@ -5,6 +6,13 @@ export class ObjectCaptionsModifier extends WorkspaceCommandSyntaxModifier {
     
     constructor(context: DevToolsExtensionContext) {
         super(context, "Add Object Captions", "addObjectCaptions");
+    }
+
+    protected getParameters(uri: vscode.Uri): any {
+        let config = vscode.workspace.getConfiguration('alOutline', uri);
+        let parameters = super.getParameters(uri);
+        parameters.sortProperties = this.sortPropertiesOnSave(uri);
+        return parameters;
     }
 
 }