+++ vscode-extension/src/alsyntaxmodifiers/removeRedundantDataClassificationModifier.ts
@@ -1,3 +1,4 @@
+import * as vscode from 'vscode';
 import { DevToolsExtensionContext } from "../devToolsExtensionContext";
 import { WorkspaceCommandSyntaxModifier } from "./workspaceCommandSyntaxModifier";
 
@@ -6,5 +7,13 @@ export class RemoveRedundantDataClassificationModifier extends WorkspaceCommandS
     constructor(context: DevToolsExtensionContext) {
         super(context, "Remove Redundant DataClassification", "removeRedundantDataClassification");
     }
+
+    protected getParameters(uri: vscode.Uri): any {
+        let config = vscode.workspace.getConfiguration('alOutline', uri);
+        let parameters = super.getParameters(uri);
+        parameters.sortProperties = this.sortPropertiesOnSave(uri);
+        return parameters;
+    }
+
 
 }